backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
psycopg2-binary>=2.9
Flask-Login>=0.5
python-dotenv>=0.21
Werkzeug>=2.0
